refactor(front): extract FurnasPage into features/furnas module
Split the 1967-line pages/FurnasPage.tsx monolith into a feature slice:
- features/furnas/FurnasPage.tsx orchestrates tabs via shared TabbedShell
- 8 section components (Home, Panorama, Metodologia, Resultados,
  Participantes, Usinas, Pesquisas, Publicacoes) hold verbatim content
- features/furnas/ui/furnasContent.ts holds Furnas-specific content styles
- shared/ui/campanha provides reusable TabbedShell + header/menu chrome
- removed dead 'banco-dados' conditional block (unreachable: menu item is
  an external link)
