+++ front/src/features/furnas/sections/PublicacoesSection.tsx
@@ -1,0 +1,207 @@
+import { JSX } from "react";
+import { Block, BlockText, BlockTitle } from "../ui/furnasContent";
+import { PageContainer } from "../../../shared/ui/campanha";
+
+export function PublicacoesSection(): JSX.Element {
+  return (
+    <>
+      <PageContainer>
+        <Block>
+          <BlockTitle>Matérias</BlockTitle>
+          <BlockText>
+            <ul>
+              <li>As Muitas Faces de uma Lagoa - Ciência Hoje, setembro de 1999</li>
+              <li>Capacitação do Setor Elétrico Brasileiro em Relação à Mudança Global do Clima</li>
+              <li>
+                Energia Renovável e Limpa: Pesquisa revela que hidrelétricas de FURNAS emitem cem
+                vezes menos gases de efeito estufa que termelétricas. Revista Furnas, junho de 2007
+              </li>
+              <li>
+                FURNAS inicia pesquisa de balanço de carbono em reservatórios - Linha Direta Nº 297,
+                14 de junho de 2003
+              </li>
+            </ul>
+          </BlockText>
+        </Block>
+
+        <Block>
+          <BlockTitle>Publicações em Revistas e Livros</BlockTitle>
+          <BlockText>
+            <ul>
+              <li>
+                Carbon gas emission from the sediments of reservoirs of different ages in central
+                Brazil.
+                <br />
+                <i>
+                  In: Global Warming and Hydroelectric Reservoirs. COPPE/UFRJ e Eletrobrás, 2005
+                </i>
+              </li>
+              <li>
+                Carbon gas cycling in the sediments of Serra da Mesa and Manso reservoirs, central
+                Brazil.
+                <br />
+                <i>
+                  Verhandlungen - Internationale Vereinigung für Theoretische und Angewandte
+                  Limnologie, 2005
+                </i>
+              </li>
+              <li>
+                Princípios físicos e químicos a serviço da limnologia - um exercício.
+                <br />
+                <i>Lições de Limnologia, São Carlos, 2005</i>
+              </li>
+              <li>
+                Caminhos do fósforo em ecossistemas aquáticos continentais.
+                <br />
+                <i>Lições de Limnologia, São Carlos, 2005</i>
+              </li>
+              <li>
+                Carbon dioxide and methane fluxes in the littoral zone of a tropical savanna
+                reservoir (Corumbá, Brazil).
+                <i>Submitted to Journal of Geophysical Research - Biogeosciences</i>
+              </li>
+              <li>
+                Photoacoustic/dynamic chamber method for measuring greenhouse gas fluxes in
+                hydroreservoirs. <i>Verhandlungen, 2005</i>
+              </li>
+              <li>Satellite ecohydrology and multifractals: INPE Relatório Técnico</li>
+              <li>
+                The use of remote sensing and automated water quality systems...{" "}
+                <i>Greenhouse gas emissions from hydropower reservoirs, COPPE/UFRJ, 2004</i>
+              </li>
+              <li>
+                Extreme event dynamics in methane ebullition fluxes...{" "}
+                <i>Geophysical Research Letters, 2006</i>
+              </li>
+              <li>
+                Long term monitoring of greenhouse gas emissions... <i>COPPE/UFRJ, 2004</i>
+              </li>
+              <li>
+                Carbon dioxide and methane emissions from hydroelectric reservoirs in Brazil.{" "}
+                <i>Global Warming and Hydroelectric Reservoirs, 2005</i>
+              </li>
+              <li>
+                Global warming and hydroelectric reservoirs. <i>COPPE/UFRJ, 2005</i>
+              </li>
+              <li>
+                Gross greenhouse gas fluxes from hydro-power reservoir compared to thermo-power
+                plants. <i>Energy Policy, 2005</i>
+              </li>
+              <li>
+                Greenhouse gases and initial findings on the carbon circulation...{" "}
+                <i>Verhandlungen, 2005</i>
+              </li>
+              <li>
+                Caminhos do carbono em ecossistemas aquáticos continentais.{" "}
+                <i>Lições de Limnologia, 2005</i>
+              </li>
+              <li>
+                Variability of carbon dioxide flux from tropical (Cerrado) hydroelectric reservoirs.{" "}
+                <i>Aquatic Sciences, 2010</i>
+              </li>
+              <li>
+                Silicon as a permanent-carbon sedimentation tracer. <i>Inland Waters, 2012</i>
+              </li>
+            </ul>
+          </BlockText>
+        </Block>
+
+        <Block>
+          <BlockTitle>Participações em Congressos</BlockTitle>
+          <BlockText>
+            <ul>
+              <li>ASLO - 2006. Victoria, Canada</li>
+              <li>
+                Greenhouse gas concentrations and diffusive flux at the sediment-water interface...
+                (Brasil)
+              </li>
+              <li>Carbon budget in two neotropical reservoirs</li>
+              <li>Management strategies to minimize bacterial methane emission...</li>
+              <li>Extreme event dynamics in methane bubbling...</li>
+              <li>Heterotrophic pathways on carbon balance in tropical reservoirs</li>
+              <li>Land use-stream carbon fluxes relationship in tropical hydro reservoir</li>
+              <li>The effect of cold fronts over the emission patterns of CO₂ and CH₄...</li>
+              <li>The fitting of Weibull PDF for surface winds...</li>
+              <li>SIL - 2004. Lahti, Finland (XXIX Congress of IAL)</li>
+              <li>Carbon gas cycling in the sediments of Serra da Mesa and Manso reservoirs...</li>
+              <li>Theoretical diffusive flux of greenhouse gases...</li>
+              <li>
+                Carbon budget in hydroelectric reservoirs of Furnas Centrais Elétricas S.A., Brazil
+              </li>
+              <li>Carbon content in the zooplankton populations of Serra da Mesa Reservoir...</li>
+              <li>Methane emission downstream of reservoirs</li>
+              <li>Preliminary results of photoacoustic/dynamic chamber technique...</li>
+              <li>Greenhouse gases and the carbon circulation in a reservoir...</li>
+              <li>SIL - 2007. Montreal, Canada (XXX Congress of IAL)</li>
+              <li>Greenhouse gas emissions from natural ecosystems and reservoirs</li>
+              <li>Carbon budget in seven Brazilian hydropower reservoirs</li>
+              <li>Greenhouse gas emissions downstream tropical hydroelectric reservoirs</li>
+              <li>
+                Greenhouse gas concentrations and diffusive flux at the sediment-water interface...
+              </li>
+              <li>Sunlight effects on diel CO₂ and CH₄ emissions...</li>
+              <li>
+                Evaluation of dissolved carbon dioxide and methane at three tropical hydroelectric
+                reservoirs
+              </li>
+              <li>Contribution of planktonic respiration to greenhouse emissions...</li>
+              <li>The importance of land use changes analysis...</li>
+              <li>Tropical reservoirs are on average 2.7 times bigger carbon sinks...</li>
+              <li>
+                Does methane from hydro-reservoirs fizz out from the water upon turbine discharge?
+              </li>
+              <li>Outros Congressos (SBRS, Simpósios e Seminários com lista completa)</li>
+            </ul>
+          </BlockText>
+        </Block>
+
+        <Block>
+          <BlockTitle>Resumos Publicados</BlockTitle>
+          <BlockText>
+            <ul>
+              <li>
+                Trophic classifications between temperate and tropical aquatic ecosystems...{" "}
+                <i>11th World Lakes Conference, 2005</i>
+              </li>
+              <li>
+                Carbon budget in hydroelectric reservoirs of FURNAS... <i>Proceedings IAL, 2005</i>
+              </li>
+              <li>
+                Concentração de oxigênio e suas implicações no metabolismo bacteriano...{" "}
+                <i>XXIII Congresso Brasileiro de Microbiologia, 2005</i>
+              </li>
+              <li>
+                Ferramentas para abrir uma caixa, ainda, nebulosa...{" "}
+                <i>X Congresso Brasileiro de Limnologia, 2005</i>
+              </li>
+              <li>
+                Carbon budget in two neotropical reservoirs...{" "}
+                <i>ASLO Aquatic Sciences Meeting, 2006</i>
+              </li>
+              <li>
+                Heterotrophic pathways on carbon balance in tropical reservoirs...{" "}
+                <i>ASLO Aquatic Sciences Meeting, 2006</i>
+              </li>
+            </ul>
+          </BlockText>
+        </Block>
+
+        <Block>
+          <BlockTitle>Monografias</BlockTitle>
+          <BlockText>
+            <ul>
+              <li>
+                Comunidade zooplanctônica de quatro reservatórios do centro-oeste do Brasil:
+                abundância e biomassa em carbono. <i>UNIRIO, 2005</i>
+              </li>
+              <li>
+                Dinâmica horizontal do fitoplâncton no reservatório de Serra da Mesa (GO)...{" "}
+                <i>UNIRIO, 2007</i>
+              </li>
+            </ul>
+          </BlockText>
+        </Block>
+      </PageContainer>
+    </>
+  );
+}
