+++ front/src/shared/ui/campanha/index.ts
@@ -1,0 +1,3 @@
+export { TabbedShell } from "./TabbedShell";
+export type { CampanhaTab } from "./TabbedShell";
+export * from "./campanhaChrome";
